Adding the DriverProcessStarting event to the .NET DriverService object
This event gives users the ability to modify the ProcessStartInfo instance
used to start the executable for the given DriverService. Possible use
cases include being able to redirect the streams for stdin, stdout and
stderr, or running the driver service executable as a different user
context than the user currently logged into the machine.