Disable log methods for .NET in the case of a spec-compliant driver
When calling driver.Manage().Logs.AvailableLogTypes for a driver that
implements the W3C WebDriver Specification (where there is no
corresponding end point), the bindings would throw a
NullReferenceException, because the command does not exist in the
command repository for spec-compliant remote ends. With this commit,
the .NET bindings will not even attempt the remote call for spec-
compliant sessions, returning empty lists instead. Fixes issue #5842.