Introduce a basic message bus
This is based on ZeroMQ, and can run either in-
process or over TCP. This should be sufficient
for our needs in GridTNG.

Also missing is a version backed by ActiveMQ and proper
JMS. That too can wait.

The aim here is to provide a pluggable mechanism for
sending messages between disparate parts of GridTNG
without introducing direct dependencies.