[grid] Introduce a basic message bus
This is based on ZeroMQ, and can run either in-
process or over TCP. This should be sufficient
for our needs in GridTNG.

There is the possibility of using ActiveMQ or JMS
at a later date, so we hide our use of ZeroMQ
behind an interface.

The aim here is to provide a pluggable mechanism for
sending messages between disparate parts of GridTNG
without introducing direct dependencies.