Allowing null value for script timeout in IE
The W3C WebDriver Specification has contradictory language regarding
the script timeout. On the one hand, it specifies that the value of
the script timeout can be `null`, indicating an indefinite timeout.
On the other hand, it specifies that when setting or getting timeouts,
the value of any timeout must be an integer between 0 and 2^53 - 1.
Since geckodriver has made the assumption that the former condition
is the correct interpretation, and the maintainers of geckodriver have
added a test to the W3C Web Platform Tests for WebDriver that expects
this interpretation, the IE driver is being modified to follow that
decision. It is the opinion of the developers of the IE driver that
this is the incorrect interpretation, but there is no recourse to have
the geckodriver maintainers revisit their decision.