Adding Firefox-specific set context method for .NET
Geckodriver/Marionette allow the user to switch the execution context
from the page being automated ("content" context) to the window
surrounding and hosting the page ("chrome" context). This commit adds
a SetContext method to FirefoxDriver to let users switch to the
desired context. The method takes an enum value to prevent switching
to invalid contexts. Note carefully that when switching out of
"content" context, not all WebDriver commands will continue to be
valid. The Selenium project does not check for the proper context
before issuing commands to geckodriver, and the user is expected
to know what context the execution is happening in once the new
SetContext method is called.