Again updated Unicode character processing for composition.
The previous change in 3.141.5.1 for using composed Unicode character
sequences on send and decomposed on retrieve was disturbingly naive,
and entirely incorrect. This revision mostly reverts that change,
with the difference of checking for "single-character" encoding
sequences when sending. The IE driver will no longer produce the same
input for some key combinations as other drivers (see some Hangul
sequences for concrete examples). The mitigating factor of this
difference in behavior, however, is that the driver should now
produce identical input and output between sending keys and reading
the text via either getting an element's text or its attributes.
Hopefully. There are likely to continue to be corner cases where use
of Unicode characters, does not entirely match up to expectations.
This is particularly so in languages where individual code points can
be combined to form other glyphs semantically or visually identical
to other individual code points.