Adding hack for handling cases where document.designMode is on
In IE, when document.designMode is turned on, the browser is put into
a state where anonymous JavaScript functions cannot be run. Obviously,
this breaks a huge portion of the driver, so the driver will now force
documents out of designMode when attempting to execute JavaScript.
There should be vanishingly few real-world cases where this has an
adverse effect, with most sites using the contentEditable attribute
instead.