[java] Implementing a new CLI option '--ext' to load extra JAR files
This option allows to specify a list of JAR files or directories separated by "path separator". Selenium will add the specified JAR files and all JAR files located in the specified directories to the classpath used to load and run commands. The '--ext' option should be used before command name.